improve semantic html, reorganize file structure
