+++ src/App.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
-import StopWatch from './components/StopWatch'
+import StopWatch from './components/StopWatch/StopWatch'
 
 export default function App() {
-    return(
-        <StopWatch/>
+    return (
+        <main>
+            <h1>StopWatch</h1>
+            <StopWatch />
+        </main>
     )
 }
